Move FakeNeoPixels into its own module
For reuse by other test modules
